requirement.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==1.1.0
langchain-core==0.3.7
google-generativeai==0.7.2
faiss-cpu==1.8.0
pypdf==4.3.1
streamlit==1.40.2
tiktoken==0.7.0
python-dotenv==1.0.1
huggingface-hub==0.23.4
sentence-transformers==3.0.1
pydantic==2.8.2
